# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shop API",
    "version": "1.2.0",
    "description": "REST API for the Shop application"
  },
  "servers": [{ "url": "http://localhost:3001" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Alex" },
                  "email": { "type": "string", "example": "alex@example.com" },
                  "password": { "type": "string", "example": "secret123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created, returns JWT token" },
          "400": {
            "description": "Validation error or email already registered"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "alex@example.com" },
                  "password": { "type": "string", "example": "secret123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns JWT token and user data" },
          "400": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout and invalidate refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged out" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "New access token" },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify OTP and activate account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp"],
                "properties": {
                  "email": { "type": "string", "example": "alex@example.com" },
                  "otp": { "type": "string", "example": "1234" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account verified, returns tokens" },
          "400": { "description": "Invalid OTP" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/auth/forgot": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request password reset OTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "example": "alex@example.com" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP sent" },
          "400": { "description": "Email is required" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/auth/verify-otp-forgot": {
      "post": {
        "tags": ["Auth"],
        "summary": "Verify password reset OTP and get reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp"],
                "properties": {
                  "email": { "type": "string", "example": "alex@example.com" },
                  "otp": { "type": "string", "example": "1234" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns resetToken" },
          "400": { "description": "Invalid OTP" }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Set new password using reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["resetToken", "password"],
                "properties": {
                  "resetToken": { "type": "string" },
                  "password": { "type": "string", "example": "newpassword123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password changed successfully" },
          "400": { "description": "Invalid or expired reset token" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user from token",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Returns current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "description": "Invalid or missing token" }
        }
      }
    },
...
    "/api/coupons": {
      "get": {
        "tags": ["Coupons"],
        "summary": "Get all coupons",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of coupons" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Coupons"],
        "summary": "Create a coupon (admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code", "discountType", "discountValue"],
                "properties": {
                  "code": { "type": "string", "example": "SAVE10" },
                  "discountType": { "type": "string", "enum": ["percent", "fixed"] },
                  "discountValue": { "type": "number", "example": 10 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Coupon created" },
          "400": { "description": "Validation error or duplicate code" },
          "403": { "description": "Admin access required" }
        }
      }
    },
    "/api/coupons/{code}": {
      "put": {
        "tags": ["Coupons"],
        "summary": "Update a coupon (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "discountType": { "type": "string", "enum": ["percent", "fixed"] },
                  "discountValue": { "type": "number", "example": 20 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Coupon updated" },
          "400": { "description": "Validation error" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Coupon not found" }
        }
      },
      "delete": {
        "tags": ["Coupons"],
        "summary": "Delete a coupon (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Coupon deleted" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Coupon not found" }
        }
      }
    },
...
  }
}